OCP working except for random cpu mem usage
