+++ nise/generators/ocp/ocp_generator.py
@@ -37,8 +37,11 @@ OCP_COLUMNS = ('report_period_start', 'report_period_end', 'pod', 'namespace',
 class OCPGenerator(AbstractGenerator):
     """Defines a abstract class for generators."""
 
-    def __init__(self, start_date, end_date):
+    def __init__(self, start_date, end_date, attributes):
         """Initialize the generator."""
+        if attributes:
+            self._nodes = attributes.get('nodes')
+
         super().__init__(start_date, end_date)
         self.apps = [self.fake.word(), self.fake.word(), self.fake.word(),  # pylint: disable=no-member
                      self.fake.word(), self.fake.word(), self.fake.word()]  # pylint: disable=no-member
@@ -62,25 +65,40 @@ class OCPGenerator(AbstractGenerator):
     def _gen_nodes(self):
         """Create nodes for report."""
         nodes = []
-        num_nodes = randint(2, 6)
-        for node_num in range(0, num_nodes):  # pylint: disable=W0612
-            memory_gig = randint(2, 8)
-            memory_bytes = memory_gig * 1024 * 1024
-            node = {'name': 'node_' + self.fake.word(),  # pylint: disable=no-member
-                    'cpu_cores': randint(2, 16),
-                    'memory_bytes': memory_bytes}
-            nodes.append(node)
+        if self._nodes:
+            for item in self._nodes:
+                memory_gig = item.get('memory_gig')
+                memory_bytes = memory_gig * 1024 * 1024
+                node = {'name': item.get('node_name'),  # pylint: disable=no-member
+                        'cpu_cores': item.get('cpu_cores'),
+                        'memory_bytes': memory_bytes,
+                        'namespaces': item.get('namespaces')}
+                nodes.append(node)
+        else:
+            num_nodes = randint(2, 6)
+            for node_num in range(0, num_nodes):  # pylint: disable=W0612
+                memory_gig = randint(2, 8)
+                memory_bytes = memory_gig * 1024 * 1024
+                node = {'name': 'node_' + self.fake.word(),  # pylint: disable=no-member
+                        'cpu_cores': randint(2, 16),
+                        'memory_bytes': memory_bytes}
+                nodes.append(node)
         return nodes
 
     def _gen_namespaces(self, nodes):
         """Create namespaces on specific nodes and keep relationship."""
         namespaces = {}
         for node in nodes:
-            num_namespaces = randint(2, 12)
-            for num_namespace in range(0, num_namespaces):  # pylint: disable=W0612
-                namespace_suffix = choice(('ci', 'qa', 'prod', 'proj', 'dev', 'staging'))
-                namespace = self.fake.word() + '_' + namespace_suffix  # pylint: disable=no-member
-                namespaces[namespace] = node
+            if node.get('namespaces'):
+                for name, _ in node.get('namespaces').items():
+                    namespace = name
+                    namespaces[namespace] = node
+            else:
+                num_namespaces = randint(2, 12)
+                for num_namespace in range(0, num_namespaces):  # pylint: disable=W0612
+                    namespace_suffix = choice(('ci', 'qa', 'prod', 'proj', 'dev', 'staging'))
+                    namespace = self.fake.word() + '_' + namespace_suffix  # pylint: disable=no-member
+                    namespaces[namespace] = node
         return namespaces
 
     def _gen_pod_labels(self):
@@ -118,27 +136,48 @@ class OCPGenerator(AbstractGenerator):
         pods = {}
         hour = 60 * 60
         for namespace, node in namespaces.items():
-            num_pods = randint(2, 20)
-            for num_namespace in range(0, num_pods):  # pylint: disable=W0612
-                pod_suffix = ''.join(choices(ascii_lowercase, k=5))
-                pod_type = choice(('build', 'deploy', pod_suffix))
-                pod = self.fake.word() + '_' + pod_type  # pylint: disable=no-member
-                cpu_cores = node.get('cpu_cores')
-                memory_bytes = node.get('memory_bytes')
-                cpu_request = round(uniform(0.02, 1.0), 5)
-                mem_request = round(uniform(250000000.0, 800000000.0), 2)
-                pods[pod] = {'namespace': namespace,
-                             'node': node.get('name'),
-                             'pod': pod,
-                             'node_capacity_cpu_cores': cpu_cores,
-                             'node_capacity_cpu_core_seconds': cpu_cores * hour,
-                             'node_capacity_memory_bytes': memory_bytes,
-                             'node_capacity_memory_byte_seconds': memory_bytes * hour,
-                             'cpu_request': cpu_request,
-                             'cpu_limit': round(uniform(cpu_request, 1.0), 5),
-                             'mem_request': mem_request,
-                             'mem_limit': round(uniform(mem_request, 800000000.0), 2),
-                             'pod_labels': self._gen_pod_labels()}
+            if node.get('namespaces'):
+                specified_pods = node.get('namespaces').get(namespace).get('pods')
+                for specified_pod in specified_pods:
+                    pod = specified_pod.get('pod_name')
+                    cpu_cores = node.get('cpu_cores')
+                    memory_bytes = node.get('memory_bytes')
+                    cpu_request = specified_pod.get('cpu_request')
+                    mem_request = specified_pod.get('mem_request')
+                    pods[pod] = {'namespace': namespace,
+                                'node': node.get('name'),
+                                'pod': pod,
+                                'node_capacity_cpu_cores': cpu_cores,
+                                'node_capacity_cpu_core_seconds': cpu_cores * hour,
+                                'node_capacity_memory_bytes': memory_bytes,
+                                'node_capacity_memory_byte_seconds': memory_bytes * hour,
+                                'cpu_request': cpu_request,
+                                'cpu_limit': specified_pod.get('cpu_limit'),
+                                'mem_request': mem_request,
+                                'mem_limit': specified_pod.get('mem_limit'),
+                                'pod_labels': specified_pod.get('labels')}
+            else:
+                num_pods = randint(2, 20)
+                for num_namespace in range(0, num_pods):  # pylint: disable=W0612
+                    pod_suffix = ''.join(choices(ascii_lowercase, k=5))
+                    pod_type = choice(('build', 'deploy', pod_suffix))
+                    pod = self.fake.word() + '_' + pod_type  # pylint: disable=no-member
+                    cpu_cores = node.get('cpu_cores')
+                    memory_bytes = node.get('memory_bytes')
+                    cpu_request = round(uniform(0.02, 1.0), 5)
+                    mem_request = round(uniform(250000000.0, 800000000.0), 2)
+                    pods[pod] = {'namespace': namespace,
+                                'node': node.get('name'),
+                                'pod': pod,
+                                'node_capacity_cpu_cores': cpu_cores,
+                                'node_capacity_cpu_core_seconds': cpu_cores * hour,
+                                'node_capacity_memory_bytes': memory_bytes,
+                                'node_capacity_memory_byte_seconds': memory_bytes * hour,
+                                'cpu_request': cpu_request,
+                                'cpu_limit': round(uniform(cpu_request, 1.0), 5),
+                                'mem_request': mem_request,
+                                'mem_limit': round(uniform(mem_request, 800000000.0), 2),
+                                'pod_labels': self._gen_pod_labels()}
         return pods
 
     def _init_data_row(self, start, end):  # noqa: C901
@@ -194,16 +233,23 @@ class OCPGenerator(AbstractGenerator):
             start = hour.get('start')
             end = hour.get('end')
             pod_count = len(self.pods)
-            num_pods = randint(2, pod_count)
-            pod_index_list = range(pod_count)
-            pod_choices = list(set(choices(pod_index_list, k=num_pods)))
-            pod_keys = list(self.pods.keys())
-            for pod_choice in pod_choices:
-                pod_name = pod_keys[pod_choice]
-                pod = deepcopy(self.pods[pod_name])
-                row = self._init_data_row(start, end)
-                row = self._update_data(row, start, end, pod=pod)
-                data.append(row)
+            if self._nodes:
+                for pod_name, _ in self.pods.items():
+                    pod = deepcopy(self.pods[pod_name])
+                    row = self._init_data_row(start, end)
+                    row = self._update_data(row, start, end, pod=pod)
+                    data.append(row)
+            else:
+                num_pods = randint(2, pod_count)
+                pod_index_list = range(pod_count)
+                pod_choices = list(set(choices(pod_index_list, k=num_pods)))
+                pod_keys = list(self.pods.keys())
+                for pod_choice in pod_choices:
+                    pod_name = pod_keys[pod_choice]
+                    pod = deepcopy(self.pods[pod_name])
+                    row = self._init_data_row(start, end)
+                    row = self._update_data(row, start, end, pod=pod)
+                    data.append(row)
         return data
 
     def generate_data(self):
