+++ ocp_static_data.yml
@@ -11,12 +11,41 @@ generators:
           namespaces:
             blah_ci:
               pods:
-                pod_name: pod_name1
-                cpu_request: 5
-                mem_request: 2
-                cpu_limit: 5
-                mem_limit: 2
-                labels:
-                  label_key1: label_value1
-                  label_key2: label_value2
+                - pod:
+                  pod_name: pod_name1
+                  cpu_request: 5
+                  mem_request: 2
+                  cpu_limit: 5
+                  mem_limit: 2
+                  labels:
+                    label_key1: label_value1
+                    label_key2: label_value2
+                - pod:
+                  pod_name: pod_name2
+                  cpu_request: 10
+                  mem_request: 4
+                  cpu_limit: 10
+                  mem_limit: 4
+                  labels:
+                    label_key3: label_value3
+                    label_key4: label_value4
             blah_qa:
+              pods:
+                - pod:
+                  pod_name: pod_name3
+                  cpu_request: 5
+                  mem_request: 2
+                  cpu_limit: 5
+                  mem_limit: 2
+                  labels:
+                    label_key1: label_value1
+                    label_key2: label_value2
+                - pod:
+                  pod_name: pod_name4
+                  cpu_request: 10
+                  mem_request: 4
+                  cpu_limit: 10
+                  mem_limit: 4
+                  labels:
+                    label_key3: label_value3
+                    label_key4: label_value4
